+++ Keysharp.Core/Internals/Window/Linux/X11/Events/XDestroyWindowEvent.cs
@@ -10,7 +10,8 @@ namespace Keysharp.Internals.Window.Linux.X11
 		internal bool send_event;
 		internal nint display;
 		internal nint xevent;
-		internal uint window;
+		internal nint window;   // Window/XID is 8 bytes on 64-bit (value happened to read correctly here as the
+		                        // last field, but keep it pointer-width for correctness and consistency).
 	}
 }
 #endif
